+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog of threedigrid-builder
 0.5.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fixed exchange_level (dpumax) for 1D2D lines connected to connection nodes
+  without a manhole. These are now the lowest of all connected channel's bank
+  levels.
 
 
 0.5.2 (2021-11-02)